Split the /cube super-command into focused MTG commands
/cube had grown to 13 subcommands spanning cube-building, card lookups,
reference and price watches. Break it into five themed top-level commands:

- /cube      → asfan, preview, generate, saved (cube building)
- /card      → lookup, rulings, combos (single-card info)
- /deck      → legality (deck analysis)
- /mtg       → set, rule (quick reference)
- /pricewatch → add, list, remove (price alerts)

Shared plumbing is extracted so each command stays small:
- AbstractMtgCommand: launchHandling + reply (the per-command boilerplate).
- MtgPoolResolver: the saved-cube/query → pool resolution shared by
  /cube preview|generate and /deck legality.
- DeckAutoComplete mirrors CubeAutoComplete for /deck's saved option.

Handler logic is moved verbatim (same embeds/services), so behaviour is
unchanged — only the command surface is reorganised. The web /cube page is
untouched (already well-factored with tabs + navbar deep-links). Tests are
split per command (CubeCommandTest trimmed; new Card/Deck/MtgReference/
PriceWatch command tests); full discord-bot suite + kover green.
