+++ discord-bot/src/main/kotlin/bot/toby/command/commands/mtg/CubeCommand.kt
@@ -4,26 +4,16 @@ import bot.toby.helpers.booleanOption
 import bot.toby.helpers.intOption
 import bot.toby.helpers.stringOption
 import common.mtg.AsFan
-import common.mtg.CardListParser
 import common.mtg.CubeAnalytics
-import common.mtg.CubeCard
-import common.mtg.DeckLegality
 import common.mtg.MtgCurrency
-import common.mtg.MtgNames
 import common.mtg.PackGenerator
-import core.command.Command.Companion.invokeDeleteOnMessageResponse
 import core.command.CommandContext
 import database.dto.guild.ConfigDto
-import database.dto.user.CardPriceWatchDto
 import database.dto.user.UserDto
 import database.service.guild.ConfigService
-import database.service.user.CardPriceWatchService
 import database.service.user.CubeListService
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import net.dv8tion.jda.api.entities.MessageEmbed
 import net.dv8tion.jda.api.interactions.commands.OptionType
 import net.dv8tion.jda.api.interactions.commands.build.OptionData
 import net.dv8tion.jda.api.interactions.commands.build.SubcommandData
@@ -33,33 +23,22 @@ import org.springframework.stereotype.Component
 import kotlin.random.Random
 
 /**
- * `/cube` — the cube-design tool from the doc, as a Discord command.
- *
- *  - `/cube asfan`    — the as-fan calculator: (type ÷ cube) × pack size.
- *  - `/cube preview`  — fetch a pool from Scryfall and show its as-fan
- *                       distribution across colours / colourless / lands.
- *  - `/cube generate` — fetch a pool, randomly draw enough cards, and deal
- *                       them into evenly-sized, as-fan-balanced packs;
- *                       the full pack lists come back as a text file.
- *  - `/cube saved`    — list the cubes the user saved on the website, each
- *                       with its card count, ready to feed back into the
- *                       `saved` option of preview/generate.
+ * `/cube` — the cube-design tools: as-fan maths, pack preview/generation, and
+ * listing your saved cubes. Card lookups, deck legality, set/keyword reference
+ * and price watches each live in their own command ([CardCommand],
+ * [DeckCommand], [MtgReferenceCommand], [PriceWatchCommand]).
  *
  * A "cube" is defined by a Scryfall search query (`set:vow`, `cube:vintage`,
- * `t:dragon`, …) — or, via the `saved` option, by one of the user's own
- * cubes saved on the website (resolved name-by-name through Scryfall).
+ * `t:dragon`, …) — or, via the `saved` option, by one of the user's own cubes
+ * saved on the website (resolved name-by-name through Scryfall).
  */
 @Component
 class CubeCommand @Autowired constructor(
-    private val fetcher: ScryfallCubeFetcher,
+    private val resolver: MtgPoolResolver,
     private val cubeListService: CubeListService,
     private val configService: ConfigService,
-    private val priceWatchService: CardPriceWatchService,
-    // Mirrors the /dnd command: IO-bound subcommands run on this dispatcher
-    // (tests pass Dispatchers.Unconfined so the launched work resolves
-    // synchronously).
-    private val dispatcher: CoroutineDispatcher = Dispatchers.IO,
-) : MtgCommand {
+    dispatcher: CoroutineDispatcher = Dispatchers.IO,
+) : AbstractMtgCommand(dispatcher) {
 
     override val name: String = "cube"
     override val description: String = "Magic: The Gathering cube tools — as-fan maths and randomised draft packs."
@@ -71,92 +50,7 @@ class CubeCommand @Autowired constructor(
             SUB_PREVIEW -> launchHandling(ctx) { handlePreview(ctx, requestingUserDto, deleteDelay) }
             SUB_GENERATE -> launchHandling(ctx) { handleGenerate(ctx, requestingUserDto) }
             SUB_SAVED -> launchHandling(ctx) { handleSaved(ctx, requestingUserDto, deleteDelay) }
-            SUB_CARD -> launchHandling(ctx) { handleCard(ctx, deleteDelay) }
-            SUB_RULINGS -> launchHandling(ctx) { handleRulings(ctx, deleteDelay) }
-            SUB_LEGALITY -> launchHandling(ctx) { handleLegality(ctx, requestingUserDto, deleteDelay) }
-            SUB_COMBOS -> launchHandling(ctx) { handleCombos(ctx, deleteDelay) }
-            SUB_SET -> launchHandling(ctx) { handleSet(ctx, deleteDelay) }
-            SUB_RULE -> handleRule(ctx, deleteDelay) // static glossary, no IO
-            SUB_WATCH_ADD -> launchHandling(ctx) { handleWatchAdd(ctx, requestingUserDto, deleteDelay) }
-            SUB_WATCH_LIST -> handleWatchList(ctx, requestingUserDto, deleteDelay) // DB read, no network
-            SUB_WATCH_REMOVE -> handleWatchRemove(ctx, requestingUserDto, deleteDelay)
-            else -> reply(ctx, CubeEmbeds.errorEmbed("Pick a subcommand: asfan, preview, generate, saved, card, rulings, legality, combos, set, rule, watch-add, watch-list or watch-remove."), deleteDelay)
-        }
-    }
-
-    /**
-     * Runs an IO-bound subcommand on [dispatcher] (the Scryfall fetch and the
-     * saved-cube DB read both block), mirroring the `/dnd` query handler, so
-     * the work never ties up the gateway or the CPU-bound default dispatcher.
-     * An unexpected failure resolves the deferred reply with an error embed
-     * rather than leaving Discord's "thinking…" spinner hanging.
-     */
-    private fun launchHandling(ctx: CommandContext, block: suspend () -> Unit) {
-        CoroutineScope(dispatcher).launch {
-            runCatching { block() }.onFailure { e ->
-                logger.error("Cube subcommand '${ctx.event.subcommandName}' failed: $e")
-                ctx.event.hook.sendMessageEmbeds(
-                    CubeEmbeds.errorEmbed("Something went wrong building that cube. Try again in a moment.")
-                ).queue({}, {})
-            }
-        }
-    }
-
-    /** A draftable pool plus a human label, any names that didn't resolve, and an optional note. */
-    private sealed interface PoolResult {
-        data class Ready(
-            val pool: List<CubeCard>,
-            val label: String,
-            val notFound: List<String> = emptyList(),
-            val note: String? = null,
-        ) : PoolResult
-        data class Failed(val message: String) : PoolResult
-    }
-
-    /**
-     * Resolves the card pool from whichever source the user gave: a saved
-     * cube (looked up for their Discord account, then resolved name-by-name
-     * via Scryfall) takes precedence over a Scryfall `query`.
-     */
-    private suspend fun resolvePool(ctx: CommandContext, requestingUserDto: UserDto): PoolResult {
-        val saved = ctx.event.stringOption(OPT_SAVED)?.trim()
-        if (!saved.isNullOrEmpty()) {
-            val dto = cubeListService.get(requestingUserDto.discordId, saved)
-                ?: return PoolResult.Failed("You have no saved cube named `$saved`. Save one on the website first.")
-            val entries = CardListParser.parse(dto.cards)
-            if (entries.isEmpty()) return PoolResult.Failed("Your saved cube `$saved` is empty.")
-            // Resolve by front face — Scryfall's collection lookup matches a
-            // single face, not the full "A // B" name; matchKeys ties the
-            // returned full-name cards back to the entries below.
-            return when (val res = fetcher.fetchByNames(entries.map { MtgNames.requestName(it.name) })) {
-                is ScryfallCubeFetcher.Result.Failure -> PoolResult.Failed(res.message)
-                is ScryfallCubeFetcher.Result.Success -> {
-                    // Index by full name AND each face so a pasted front-face
-                    // name (e.g. "Archangel Avacyn") matches a transform card's
-                    // full name ("Archangel Avacyn // Avacyn, the Purifier").
-                    val byName = HashMap<String, CubeCard>()
-                    res.cards.forEach { card ->
-                        MtgNames.matchKeys(card.name).forEach { key -> byName.putIfAbsent(key, card) }
-                    }
-                    val pool = entries.flatMap { entry ->
-                        byName[MtgNames.lookupKey(entry.name)]?.let { card -> List(entry.count) { card } } ?: emptyList()
-                    }
-                    val notFound = entries.map { it.name }
-                        .filter { byName[MtgNames.lookupKey(it)] == null }
-                        .distinct()
-                    if (pool.isEmpty()) PoolResult.Failed("None of `$saved`'s cards matched Scryfall.")
-                    else PoolResult.Ready(pool, "saved cube \"$saved\"", notFound, capNote(res.capped))
-                }
-            }
-        }
-
-        val query = ctx.event.stringOption(OPT_QUERY)?.trim()
-        if (query.isNullOrEmpty()) {
-            return PoolResult.Failed("Give me a Scryfall `query`, or the `saved` name of one of your cubes.")
-        }
-        return when (val res = fetcher.fetch(query)) {
-            is ScryfallCubeFetcher.Result.Failure -> PoolResult.Failed(res.message)
-            is ScryfallCubeFetcher.Result.Success -> PoolResult.Ready(res.cards, query, note = capNote(res.capped))
+            else -> reply(ctx, CubeEmbeds.errorEmbed("Pick a subcommand: asfan, preview, generate or saved."), deleteDelay)
         }
     }
 
@@ -170,10 +64,8 @@ class CubeCommand @Autowired constructor(
             configService.getConfigByName(ConfigDto.Configurations.CUBE_CURRENCY.configValue, ctx.guild.id)?.value
         ) ?: MtgCurrency.DEFAULT
 
-    /** A user-facing note when the pool was truncated to the Scryfall fetch ceiling. */
-    private fun capNote(capped: Boolean): String? =
-        if (capped) "Matched more than ${ScryfallCubeFetcher.DEFAULT_MAX_CARDS} cards; only the first ${ScryfallCubeFetcher.DEFAULT_MAX_CARDS} were used."
-        else null
+    private suspend fun resolvePool(ctx: CommandContext, requestingUserDto: UserDto) =
+        resolver.resolve(ctx.event.stringOption(OPT_SAVED), ctx.event.stringOption(OPT_QUERY), requestingUserDto.discordId)
 
     private fun handleAsFan(ctx: CommandContext, deleteDelay: Int) {
         val total = ctx.event.intOption(OPT_TOTAL, 0)
@@ -191,8 +83,8 @@ class CubeCommand @Autowired constructor(
     private suspend fun handlePreview(ctx: CommandContext, requestingUserDto: UserDto, deleteDelay: Int) {
         val packSize = ctx.event.intOption(OPT_PACK_SIZE, DEFAULT_PACK_SIZE)
         when (val resolved = resolvePool(ctx, requestingUserDto)) {
-            is PoolResult.Failed -> reply(ctx, CubeEmbeds.errorEmbed(resolved.message), deleteDelay)
-            is PoolResult.Ready -> {
+            is MtgPoolResolver.PoolResult.Failed -> reply(ctx, CubeEmbeds.errorEmbed(resolved.message), deleteDelay)
+            is MtgPoolResolver.PoolResult.Ready -> {
                 val pool = resolved.pool
                 val currency = currencyFor(ctx)
                 val embed = CubeEmbeds.previewEmbed(
@@ -218,8 +110,8 @@ class CubeCommand @Autowired constructor(
         val balanced = ctx.event.booleanOption(OPT_BALANCED, true)
 
         when (val resolved = resolvePool(ctx, requestingUserDto)) {
-            is PoolResult.Failed -> ctx.event.hook.sendMessageEmbeds(CubeEmbeds.errorEmbed(resolved.message)).queue()
-            is PoolResult.Ready -> {
+            is MtgPoolResolver.PoolResult.Failed -> ctx.event.hook.sendMessageEmbeds(CubeEmbeds.errorEmbed(resolved.message)).queue()
+            is MtgPoolResolver.PoolResult.Ready -> {
                 val pool = resolved.pool
                 when (val packs = PackGenerator(Random.Default).generate(pool, packCount, packSize, balanced)) {
                     is PackGenerator.Result.Failure -> ctx.event.hook.sendMessageEmbeds(CubeEmbeds.errorEmbed(packs.reason)).queue()
@@ -253,157 +145,6 @@ class CubeCommand @Autowired constructor(
         reply(ctx, CubeEmbeds.savedCubesEmbed(saved), deleteDelay)
     }
 
-    /** Looks up a single card by name on Scryfall and shows its image + facts. */
-    private suspend fun handleCard(ctx: CommandContext, deleteDelay: Int) {
-        val name = ctx.event.stringOption(OPT_NAME)?.trim()
-        if (name.isNullOrEmpty()) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a card `name` to look up."), deleteDelay)
-            return
-        }
-        when (val res = fetcher.fetchByNames(listOf(MtgNames.requestName(name)))) {
-            is ScryfallCubeFetcher.Result.Failure ->
-                reply(ctx, CubeEmbeds.errorEmbed("Couldn't find a card named `$name`."), deleteDelay)
-            is ScryfallCubeFetcher.Result.Success -> {
-                // Prefer the card whose full/face name matches what was typed.
-                val card = res.cards.firstOrNull { MtgNames.matchKeys(it.name).contains(MtgNames.lookupKey(name)) }
-                    ?: res.cards.firstOrNull()
-                if (card == null) reply(ctx, CubeEmbeds.errorEmbed("Couldn't find a card named `$name`."), deleteDelay)
-                else reply(ctx, CubeEmbeds.cardEmbed(card), deleteDelay)
-            }
-        }
-    }
-
-    /** Looks up a single card's official rulings on Scryfall. */
-    private suspend fun handleRulings(ctx: CommandContext, deleteDelay: Int) {
-        val name = ctx.event.stringOption(OPT_NAME)?.trim()
-        if (name.isNullOrEmpty()) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a card `name` to look up rulings for."), deleteDelay)
-            return
-        }
-        when (val rulings = fetcher.fetchRulings(name)) {
-            null -> reply(ctx, CubeEmbeds.errorEmbed("Couldn't find a card named `$name`."), deleteDelay)
-            else -> reply(ctx, CubeEmbeds.rulingsEmbed(rulings), deleteDelay)
-        }
-    }
-
-    /** Looks up the combos a single card appears in, via Commander Spellbook. */
-    private suspend fun handleCombos(ctx: CommandContext, deleteDelay: Int) {
-        val name = ctx.event.stringOption(OPT_NAME)?.trim()
-        if (name.isNullOrEmpty()) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a card `name` to find combos for."), deleteDelay)
-            return
-        }
-        when (val combos = fetcher.fetchCombos(name)) {
-            null -> reply(ctx, CubeEmbeds.errorEmbed("Couldn't reach Commander Spellbook. Try again later."), deleteDelay)
-            else -> reply(ctx, CubeEmbeds.combosEmbed(combos), deleteDelay)
-        }
-    }
-
-    /** Looks up a Magic set by its code on Scryfall. */
-    private suspend fun handleSet(ctx: CommandContext, deleteDelay: Int) {
-        val code = ctx.event.stringOption(OPT_CODE)?.trim()
-        if (code.isNullOrEmpty()) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a set `code` (e.g. vow, mh3)."), deleteDelay)
-            return
-        }
-        when (val set = fetcher.fetchSet(code)) {
-            null -> reply(ctx, CubeEmbeds.errorEmbed("Couldn't find a set with code `$code`."), deleteDelay)
-            else -> reply(ctx, CubeEmbeds.setEmbed(set), deleteDelay)
-        }
-    }
-
-    /** Looks up a keyword's reminder text in the built-in glossary (no network). */
-    private fun handleRule(ctx: CommandContext, deleteDelay: Int) {
-        val term = ctx.event.stringOption(OPT_TERM)?.trim()
-        if (term.isNullOrEmpty()) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a keyword (e.g. trample, deathtouch)."), deleteDelay)
-            return
-        }
-        when (val found = common.mtg.MtgGlossary.lookup(term)) {
-            null -> reply(ctx, CubeEmbeds.errorEmbed("No glossary entry for `$term`. Try an evergreen keyword like trample or flying."), deleteDelay)
-            else -> reply(ctx, CubeEmbeds.ruleEmbed(found), deleteDelay)
-        }
-    }
-
-    /** Adds a card price watch: resolves the card, captures its current price, persists. */
-    private suspend fun handleWatchAdd(ctx: CommandContext, requestingUserDto: UserDto, deleteDelay: Int) {
-        val name = ctx.event.stringOption(OPT_NAME)?.trim()
-        val directionKey = ctx.event.stringOption(OPT_DIRECTION)?.trim()?.uppercase()
-        val price = ctx.event.getOption(OPT_PRICE)?.asDouble
-        val currency = MtgCurrency.fromCode(ctx.event.stringOption(OPT_CURRENCY)) ?: MtgCurrency.DEFAULT
-        if (name.isNullOrEmpty()) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a card `name` to watch."), deleteDelay); return
-        }
-        val direction = directionKey?.let { runCatching { CardPriceWatchDto.Direction.valueOf(it) }.getOrNull() }
-        if (direction == null) {
-            reply(ctx, CubeEmbeds.errorEmbed("Pick a `direction`: below or above."), deleteDelay); return
-        }
-        if (price == null || price <= 0.0) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me a positive target `price`."), deleteDelay); return
-        }
-        val card = fetcher.fetchNamed(name)
-        if (card == null) {
-            reply(ctx, CubeEmbeds.errorEmbed("Couldn't find a card named `$name`."), deleteDelay); return
-        }
-        val currentPrice = card.price(currency)?.toDoubleOrNull()
-        val created = priceWatchService.create(
-            discordId = requestingUserDto.discordId,
-            guildId = ctx.guild.idLong,
-            cardName = card.name,
-            currency = currency.code,
-            direction = direction,
-            threshold = price,
-            priceAtCreation = currentPrice,
-        )
-        if (created == null) {
-            reply(ctx, CubeEmbeds.errorEmbed("You've hit the watch limit (${priceWatchService.maxPerUser}). Remove one with `/cube watch-remove` first."), deleteDelay)
-        } else {
-            reply(ctx, CubeEmbeds.watchAddedEmbed(created, card, currency, currentPrice), deleteDelay)
-        }
-    }
-
-    /** Lists the caller's card price watches. */
-    private fun handleWatchList(ctx: CommandContext, requestingUserDto: UserDto, deleteDelay: Int) {
-        val watches = priceWatchService.listForUser(requestingUserDto.discordId)
-        reply(ctx, CubeEmbeds.watchListEmbed(watches), deleteDelay)
-    }
-
-    /** Removes one of the caller's watches by id (ownership-checked). */
-    private fun handleWatchRemove(ctx: CommandContext, requestingUserDto: UserDto, deleteDelay: Int) {
-        val id = ctx.event.getOption(OPT_WATCH_ID)?.asLong
-        if (id == null) {
-            reply(ctx, CubeEmbeds.errorEmbed("Give me the watch `id` to remove (see `/cube watch-list`)."), deleteDelay); return
-        }
-        val removed = priceWatchService.remove(id, requestingUserDto.discordId)
-        reply(
-            ctx,
-            if (removed) CubeEmbeds.watchRemovedEmbed(id)
-            else CubeEmbeds.errorEmbed("No watch #$id of yours to remove."),
-            deleteDelay,
-        )
-    }
-
-    /** Checks a saved cube / queried pool against a format's banned & legality list. */
-    private suspend fun handleLegality(ctx: CommandContext, requestingUserDto: UserDto, deleteDelay: Int) {
-        val formatKey = ctx.event.stringOption(OPT_FORMAT)?.trim()?.lowercase()
-        val format = CubeCard.FORMATS.firstOrNull { it.first == formatKey }
-        if (format == null) {
-            reply(ctx, CubeEmbeds.errorEmbed("Pick a `format` to check against (e.g. Modern, Commander)."), deleteDelay)
-            return
-        }
-        when (val resolved = resolvePool(ctx, requestingUserDto)) {
-            is PoolResult.Failed -> reply(ctx, CubeEmbeds.errorEmbed(resolved.message), deleteDelay)
-            is PoolResult.Ready -> {
-                val report = DeckLegality.check(resolved.pool, format.first)
-                reply(ctx, CubeEmbeds.legalityEmbed(report, format.second, resolved.label, resolved.notFound), deleteDelay)
-            }
-        }
-    }
-
-    private fun reply(ctx: CommandContext, embed: MessageEmbed, deleteDelay: Int) {
-        ctx.event.hook.sendMessageEmbeds(embed).queue(invokeDeleteOnMessageResponse(deleteDelay))
-    }
-
     override val subCommands: List<SubcommandData> = listOf(
         SubcommandData(SUB_ASFAN, "Expected copies of a card type per booster: (type ÷ cube) × pack size.")
             .addOptions(
@@ -434,37 +175,6 @@ class CubeCommand @Autowired constructor(
                 OptionData(OptionType.BOOLEAN, OPT_BALANCED, "Level colours/lands across packs (default true).", false),
             ),
         SubcommandData(SUB_SAVED, "List the cubes you've saved on the website."),
-        SubcommandData(SUB_CARD, "Look up a single Magic card by name.")
-            .addOptions(OptionData(OptionType.STRING, OPT_NAME, "The card's name (e.g. Lightning Bolt).", true)),
-        SubcommandData(SUB_RULINGS, "Look up a Magic card's official rulings by name.")
-            .addOptions(OptionData(OptionType.STRING, OPT_NAME, "The card's name (e.g. Doubling Season).", true)),
-        SubcommandData(SUB_LEGALITY, "Check whether a saved cube/deck is legal in a format (banned & not-legal cards).")
-            .addOptions(
-                OptionData(OptionType.STRING, OPT_FORMAT, "Format to check against.", true)
-                    .addChoices(CubeCard.FORMATS.map { net.dv8tion.jda.api.interactions.commands.Command.Choice(it.second, it.first) }),
-                OptionData(OptionType.STRING, OPT_SAVED, "Name of a deck/cube you saved on the website.", false)
-                    .setAutoComplete(true),
-                OptionData(OptionType.STRING, OPT_QUERY, "Or a Scryfall search to check instead (e.g. set:mh3).", false),
-            ),
-        SubcommandData(SUB_COMBOS, "Find the combos a Magic card is part of (Commander Spellbook).")
-            .addOptions(OptionData(OptionType.STRING, OPT_NAME, "The card's name (e.g. Thassa's Oracle).", true)),
-        SubcommandData(SUB_SET, "Look up a Magic set by its code (release date, card count).")
-            .addOptions(OptionData(OptionType.STRING, OPT_CODE, "The set code (e.g. vow, mh3).", true)),
-        SubcommandData(SUB_RULE, "Look up a Magic keyword's reminder text (e.g. trample, flying).")
-            .addOptions(OptionData(OptionType.STRING, OPT_TERM, "The keyword to explain.", true)),
-        SubcommandData(SUB_WATCH_ADD, "Get DM'd when a card's price crosses your target.")
-            .addOptions(
-                OptionData(OptionType.STRING, OPT_NAME, "The card to watch (e.g. Ragavan).", true),
-                OptionData(OptionType.STRING, OPT_DIRECTION, "Alert when the price goes…", true)
-                    .addChoice("below", CardPriceWatchDto.Direction.BELOW.name)
-                    .addChoice("above", CardPriceWatchDto.Direction.ABOVE.name),
-                OptionData(OptionType.NUMBER, OPT_PRICE, "Target price.", true).setMinValue(0.0),
-                OptionData(OptionType.STRING, OPT_CURRENCY, "Currency (default USD).", false)
-                    .addChoices(MtgCurrency.entries.map { net.dv8tion.jda.api.interactions.commands.Command.Choice(it.display, it.code) }),
-            ),
-        SubcommandData(SUB_WATCH_LIST, "List your card price watches."),
-        SubcommandData(SUB_WATCH_REMOVE, "Remove one of your card price watches.")
-            .addOptions(OptionData(OptionType.INTEGER, OPT_WATCH_ID, "The watch id (from /cube watch-list).", true).setMinValue(1)),
     )
 
     companion object {
@@ -472,25 +182,8 @@ class CubeCommand @Autowired constructor(
         const val SUB_PREVIEW = "preview"
         const val SUB_GENERATE = "generate"
         const val SUB_SAVED = "saved"
-        const val SUB_CARD = "card"
-        const val SUB_RULINGS = "rulings"
-        const val SUB_LEGALITY = "legality"
-        const val SUB_COMBOS = "combos"
-        const val SUB_SET = "set"
-        const val SUB_RULE = "rule"
-        const val SUB_WATCH_ADD = "watch-add"
-        const val SUB_WATCH_LIST = "watch-list"
-        const val SUB_WATCH_REMOVE = "watch-remove"
 
         const val OPT_TOTAL = "total"
-        const val OPT_NAME = "name"
-        const val OPT_FORMAT = "format"
-        const val OPT_CODE = "code"
-        const val OPT_TERM = "term"
-        const val OPT_DIRECTION = "direction"
-        const val OPT_PRICE = "price"
-        const val OPT_CURRENCY = "currency"
-        const val OPT_WATCH_ID = "id"
         const val OPT_CUBE_SIZE = "cube-size"
         const val OPT_PACK_SIZE = "pack-size"
         const val OPT_QUERY = "query"
